+++ sdk/appcomplianceautomation/arm-appcomplianceautomation/src/operations/operations.ts
@@ -35,7 +35,7 @@ export class OperationsImpl implements Operations {
   }
 
   /**
-   * Lists all of the available REST API operations of the Microsoft.AppComplianceAutomation provider.
+   * List the operations for the provider
    * @param options The options parameters.
    */
   public list(
@@ -89,7 +89,7 @@ export class OperationsImpl implements Operations {
   }
 
   /**
-   * Lists all of the available REST API operations of the Microsoft.AppComplianceAutomation provider.
+   * List the operations for the provider
    * @param options The options parameters.
    */
   private _list(
@@ -143,7 +143,6 @@ const listNextOperationSpec: coreClient.OperationSpec = {
       bodyMapper: Mappers.ErrorResponse
     }
   },
-  queryParameters: [Parameters.apiVersion],
   urlParameters: [Parameters.$host, Parameters.nextLink],
   headerParameters: [Parameters.accept],
   serializer